+++ src/app/admin/admin.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { ReactiveFormsModule } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 import { AdminRoutingModule } from './admin-routing.module';
 
@@ -13,6 +13,7 @@ import { components } from './components';
   ],
   imports: [
     CommonModule,
+    FormsModule,
     ReactiveFormsModule,
 
     AdminRoutingModule