+++ src/app/core/guards/auth.guard.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { CanLoad, Route, Router, UrlSegment } from '@angular/router';
+import { ActivatedRouteSnapshot, CanActivate, CanLoad, Route, Router, RouterStateSnapshot, UrlSegment } from '@angular/router';
 
 import { Observable } from 'rxjs';
 
@@ -9,23 +9,29 @@ import { AuthService } from '../../auth/services';
 @Injectable({
   providedIn: CoreModule
 })
-export class AuthGuard implements CanLoad {
+export class AuthGuard implements CanLoad, CanActivate {
   constructor(
     private router: Router,
     private authService: AuthService,
   ) {
   }
 
+  canActivate(
+    route: ActivatedRouteSnapshot,
+    state: RouterStateSnapshot): Observable<boolean> | Promise<boolean> | boolean  {
+    return this.isUserAuthorized(state.url);
+  }
+
   canLoad(
     route: Route,
     segments: UrlSegment[]
   ): Observable<boolean> | Promise<boolean> | boolean {
-    return this.isUserAuthorized(route);
+    return this.isUserAuthorized(route.path);
   }
 
-  private isUserAuthorized(route: Route): boolean {
+  private isUserAuthorized(path: string): boolean {
     const authorized = this.authService.isLogin();
-    const isLoginPage = route.path.indexOf('auth') !== -1;
+    const isLoginPage = path.indexOf('auth') !== -1;
 
 
     if (!authorized && !isLoginPage) {