+++ src/app/core/services/api-config.service.ts
@@ -2,4 +2,5 @@ export class ApiConfig {
   static BASE_URL = 'http://localhost:3000/';
   static PRODUCTS_URL = `${ApiConfig.BASE_URL}products/`;
   static LOGIN_URL = `${ApiConfig.BASE_URL}login/`;
+  static ORDERS_URL = `${ApiConfig.BASE_URL}orders/`;
 }