+++ src/app/core/store/app.state.ts
@@ -1,0 +1,9 @@
+import { OrdersState } from './orders';
+import { ProductsState } from './products';
+import { RouterState } from './router';
+
+export interface AppState {
+  router: RouterState;
+  orders: OrdersState;
+  products: ProductsState;
+}