+++ src/app/core/store/core-store.module.ts
@@ -1,0 +1,24 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { StoreModule } from '@ngrx/store';
+import { EffectsModule } from '@ngrx/effects';
+import { StoreRouterConnectingModule } from '@ngrx/router-store';
+
+import { RouterEffects, routerReducers, RouterStateSerializerProvider } from './router';
+
+
+@NgModule({
+  declarations: [],
+  imports: [
+    CommonModule,
+    StoreModule.forRoot(routerReducers),
+    EffectsModule.forRoot([RouterEffects]),
+    StoreRouterConnectingModule.forRoot(),
+  ],
+  providers: [
+    RouterStateSerializerProvider,
+  ]
+})
+export class CoreStoreModule { }
+