+++ src/app/core/store/index.ts
@@ -1,0 +1,6 @@
+export * from './app.state';
+export * from './orders';
+export * from './router';
+export * from './products';
+
+export { CoreStoreModule } from './core-store.module';