+++ src/app/core/store/products/products.state.ts
@@ -1,0 +1,16 @@
+import { ProductModel } from '../../../products';
+
+export interface ProductsState {
+  data: ReadonlyArray<ProductModel>;
+  readonly loading: boolean;
+  readonly loaded: boolean;
+  readonly error: Error | string;
+}
+
+export const initialProductsState: ProductsState = {
+  data: [],
+  loading: false,
+  loaded: false,
+  error: null
+};
+