+++ src/app/products/components/product-list/product-list.component.ts
@@ -1,11 +1,14 @@
 import { Component, OnInit } from '@angular/core';
 
 import { Observable } from 'rxjs';
+import { Store } from '@ngrx/store';
 
 import { ProductModel } from '../../models';
 
 import { ProductsService } from '../../services';
 import { CartService } from '../../../cart/services';
+import { AppState, loadProducts, getProductsData, getProductsLoading } from '../../../core/store';
+
 
 @Component({
   selector: 'sh-product-list',
@@ -14,14 +17,19 @@ import { CartService } from '../../../cart/services';
 })
 export class ProductListComponent implements OnInit {
   products$: Observable<ProductModel[]>;
+  loading$: Observable<boolean>;
 
   constructor(
+    private store: Store<AppState>,
     private productsService: ProductsService,
     private cartService: CartService
   ) { }
 
   ngOnInit() {
-    this.products$ = this.productsService.getProducts$();
+    this.store.dispatch(loadProducts());
+
+    this.products$ = this.store.select(getProductsData);
+    this.loading$ = this.store.select(getProductsLoading);
   }
 
   onAddProductToCart(product: ProductModel): void {
@@ -29,5 +37,4 @@ export class ProductListComponent implements OnInit {
 
     this.cartService.addProduct(p);
   }
-
 }